pyproject.toml
[project]
name = "nonebot-plugin-nps"
version = "0.1.0"
description = "an unofficial plugin store for nonebot"
authors = [
    { name = "snowykami", email = "snowykami@outlook.com" },
]
dependencies = [
    "nonebot2>=2.3.3",
    "pydantic>=2.9.2",
    "aiohttp>=3.10.5",
    "aiofiles>=24.1.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.nonebot]
plugins = ["nonebot_plugin_nps"]

[tool.pdm.dev-dependencies]
dev = [
    "nb-cli>=1.4.2",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]
